requirements.txt
django==4.1.6
psycopg2==2.9.3
djangorestframework
